[Refactor] 공통 컴포넌트에서 투표 cell로 분리
