+++ TwoHoSun/Presentation/Review/ReviewDetailView.swift
@@ -111,11 +111,7 @@ extension ReviewDetailView {
             Button {
                 // TODO: 투표 게시글 상세 조회로 이동
             } label: {
-                // TODO: VoteCardCell refactoring
-
-//                VoteCardCell(cellType: .simple,
-//                             progressType: .closed,
-//                             data: data)
+                VoteCardCell(cellType: .standard, data: data)
             }
         }
     }
